refactor: encapsulated move data
