+++ src/main/scala/it/scalalearn/goosegame/movelogic/Move.scala
@@ -1,0 +1,3 @@
+package it.scalalearn.goosegame.movelogic
+
+case class Move(name: String, previousSquare: Int, startSquare: Int, dice: List[Int])
